- Added v2 as current main API version - Completed task object - Added the MIT license
