+++ freshservice/v2/models/exceptions.py
